+++ app/controllers/policies_controller.rb
@@ -15,6 +15,13 @@ class PoliciesController < RestController
   # is provided in the request.
   set_default_content_type_for_path(%r{^/policies}, 'application/x-yaml')
 
+  # To avoid unexpected behavior, we check annotation keys for known policy attributes and ouput a warning if there is a match.
+  KNOWN_POLICY_ATTRIBUTES = %w[
+      id owner body user annotations restricted_to permit deny role privileges resource
+      grant revoke member host-factory layers layer variable kind mime_type delete record
+      group host webservice
+    ].freeze
+
   def get
     action = :read
 
@@ -191,8 +198,10 @@ class PoliciesController < RestController
     # Success
     audit_success(policy_result.policy_version)
 
+    response = policy_mode.report(policy_result)
+    response[:warnings] = policy_result.warnings if policy_result.warnings.present?
     render(
-      json: policy_mode.report(policy_result),
+      json: response,
       status: strategy_type == :orchestration ? :created : :ok
     )
 
@@ -330,6 +339,17 @@ class PoliciesController < RestController
       root_policy: is_root
     )
     policy_result.policy_parse = (parse)
+
+    annotation_names = parse.records.flat_map do |rec|
+      rec.respond_to?(:annotations) && rec.annotations ? rec.annotations.keys : []
+    end.uniq
+    conflicts = annotation_names & KNOWN_POLICY_ATTRIBUTES
+    if conflicts.any?
+      policy_result.warnings ||= []
+      conflicts.each do |conflict|
+        policy_result.warnings << "Annotation '#{conflict}' matches a known policy attribute. This annotation will not be treated as a standard attribute and may not have the intended effect."
+      end
+    end
   end
 
   def publish_event