+++ app/controllers/secrets_controller.rb
@@ -13,6 +13,8 @@ class SecretsController < RestController
 
     value = request.raw_post
 
+    validate_public_key(value) if params[:kind] == 'public_key'
+
     raise ArgumentError, "'value' may not be empty" if value.blank?
 
     Secret.create(resource_id: resource.id, value: value)
@@ -146,6 +148,12 @@ class SecretsController < RestController
   end
 
   private
+
+  def validate_public_key(value)
+    unless value.match?(/\Assh-(rsa|ed25519|ecdsa-[a-z0-9-]+) [A-Za-z0-9+\/=]+ ?.*\z/)
+      raise ArgumentError, "Invalid public key format"
+    end
+  end
 
   def variable_ids
     return @variable_ids if @variable_ids