+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Nothing should go in this section, please add to the latest unreleased version
   (and update the corresponding date), or add a new version.
 
+## [1.21.5] - 2025-06-30
+### Added
+- Allow conjur administrator to enable additional signed headers for IAM authenticator. CNJR-10217
+
 ## [1.21.4] - 2025-01-13
 ### Security
 - Improve headers handling in AWS IAM authenticator. CONJSE-2023